#41 split bakery_mock tests
